FAC-644: Add Existence Test to Implementation of DeleteMaturityLevelPort

+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/out/persistence/maturitylevel/MaturityLevelPersistenceJpaAdapter.java
@@ -1,6 +1,7 @@
 package org.flickit.assessment.kit.adapter.out.persistence.maturitylevel;
 
 import lombok.RequiredArgsConstructor;
+import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.data.jpa.kit.levelcompetence.LevelCompetenceJpaEntity;
 import org.flickit.assessment.data.jpa.kit.levelcompetence.LevelCompetenceJpaRepository;
 import org.flickit.assessment.data.jpa.kit.maturitylevel.MaturityLevelJpaEntity;
@@ -19,6 +20,7 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 
 import static org.flickit.assessment.kit.adapter.out.persistence.maturitylevel.MaturityLevelMapper.mapToJpaEntityToPersist;
+import static org.flickit.assessment.kit.common.ErrorMessageKey.MATURITY_LEVEL_ID_NOT_FOUND;
 
 @Component
 @RequiredArgsConstructor
@@ -39,6 +41,8 @@ public class MaturityLevelPersistenceJpaAdapter implements
 
     @Override
     public void delete(Long id, Long kitVersionId) {
+        if(!repository.existsByIdAndKitVersionId(id, kitVersionId))
+            throw new ResourceNotFoundException(MATURITY_LEVEL_ID_NOT_FOUND);
         repository.deleteByIdAndKitVersionId(id, kitVersionId);
     }
